after further reflection, remove the direct use of Link.
redux actions require that this kind of stuff never be used.  Thus,
instead of using <Link>, we will simply map the onClick to a dispatch
of a route change event, and use a reducer to handle that (or
middleware, if that is the better approach)

Also, simplify the MenuItem, we don’t need a special case for no link.